move Create to Repetitive namespace
